Chore: Refactor side-menu-content to TS and React Hooks
