+++ packages/app-mobile/utils/types.ts
@@ -1,0 +1,9 @@
+import { State } from '@joplin/lib/reducer';
+
+export interface AppState extends State {
+	sideMenuOpenPercent: number;
+	isOnMobileData: boolean;
+	route: any;
+	smartFilterId: string;
+	noteSideMenuOptions: any;
+}
